Move demo reel player from utils
