+++ packages/sfpowerscripts-cli/src/SfpowerscriptsCommand.ts
@@ -3,7 +3,7 @@ import { OutputFlags } from "@oclif/parser";
 import SFPStatsSender from "@dxatscale/sfpowerscripts.core/lib/utils/SFPStatsSender"
 import * as rimraf from "rimraf";
 import ProjectValidation from "./ProjectValidation";
-import DemoReelPlayer from "./utils/demoReelPlayer";
+import DemoReelPlayer from "./impl/demoreelplayer/DemoReelPlayer";
 import { fs } from "@salesforce/core";
 
 /**
